chore: consolidate tests
